requirements.txt
# Core dependencies
crewai>=0.11.0
crewai-tools>=0.2.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# OpenAI integration
openai>=1.0.0
langchain-openai>=0.0.5

# Telegram bot (Milestone 2)
python-telegram-bot>=22.3

# CLI utilities
click>=8.1.0
rich>=13.0.0

# Data handling
pandas>=2.0.0
